requirements.txt
opencv_python==4.5.3.56
six==1.15.0
numpy==1.19.3
object_detection-0.1==0.1
matplotlib==3.1.3
tensorflow==2.5.0
object_detection==0.0.3
Pillow==8.3.1
protobuf==3.17.3
